add CharSequenceUtil
