+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * 【json   】     JSONStrFormater改为JSONStrFormatter
 * 【dfa    】     增加FoundWord（pr#1290@Github）
 * 【core   】     增加Segment（pr#1290@Github）
+* 【core   】     增加CharSequenceUtil
 
 ### Bug修复
 * 【cache  】     修复Cache中get重复misCount计数问题（issue#1281@Github）
