+++ hutool-core/src/main/java/cn/hutool/core/util/XmlUtil.java
@@ -66,6 +66,36 @@ import java.util.Map;
 public class XmlUtil {
 
 	/**
+	 * 字符串常量：XML 空格转义 {@code "&nbsp;" -> " "}
+	 */
+	public static final String NBSP = "&nbsp;";
+
+	/**
+	 * 字符串常量：XML And 符转义 {@code "&amp;" -> "&"}
+	 */
+	public static final String AMP = "&amp;";
+
+	/**
+	 * 字符串常量：XML 双引号转义 {@code "&quot;" -> "\""}
+	 */
+	public static final String QUOTE = "&quot;";
+
+	/**
+	 * 字符串常量：XML 单引号转义 {@code "&apos" -> "'"}
+	 */
+	public static final String APOS = "&apos;";
+
+	/**
+	 * 字符串常量：XML 小于号转义 {@code "&lt;" -> "<"}
+	 */
+	public static final String LT = "&lt;";
+
+	/**
+	 * 字符串常量：XML 大于号转义 {@code "&gt;" -> ">"}
+	 */
+	public static final String GT = "&gt;";
+
+	/**
 	 * 在XML中无效的字符 正则
 	 */
 	public static final String INVALID_REGEX = "[\\x00-\\x08\\x0b-\\x0c\\x0e-\\x1f]";
