+++ hutool-extra/src/main/java/cn/hutool/extra/spring/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Spring相关工具封装
+ * 
+ * @author looly
+ *
+ */
+package cn.hutool.extra.spring;
