+++ hutool-extra/src/test/java/cn/hutool/extra/spring/EnableSpringUtilTest.java
@@ -10,15 +10,15 @@ import org.springframework.test.context.junit4.SpringRunner;
  * @author sidian
  */
 @RunWith(SpringRunner.class)
-@SpringBootTest(classes = EnableSprintUtilTest.class)
+@SpringBootTest(classes = EnableSpringUtilTest.class)
 @EnableSpringUtil
-public class EnableSprintUtilTest {
+public class EnableSpringUtilTest {
 
-    @Test
-    public void test() {
-        // 使用@EnableSpringUtil注解后, 能获取上下文
-        Assert.assertNotNull(SpringUtil.getApplicationContext());
-        // 不使用时, 为null
+	@Test
+	public void test() {
+		// 使用@EnableSpringUtil注解后, 能获取上下文
+		Assert.assertNotNull(SpringUtil.getApplicationContext());
+		// 不使用时, 为null
 //        Assert.assertNull(SpringUtil.getApplicationContext());
-    }
+	}
 }
